requirements.txt
biopython==1.81
matplotlib==3.9.2
matplotlib-inline==0.1.7
numpy
pandas
plotly==5.24.1
streamlit